MOTOR-787 Drop support for Python 3.5
