+++ .evergreen/config.yml
@@ -760,18 +760,10 @@ axes:
   - id: tox-env
     display_name: "Tox Env"
     values:
-      - id: "tornado5-pypy35"
-        variables:
-           TOX_ENV: "tornado5-pypy35"
-           PYTHON_BINARY: "/opt/python/pypy3.5/bin/pypy3"
       - id: "tornado5-pypy36"
         variables:
            TOX_ENV: "tornado5-pypy36"
            PYTHON_BINARY: "/opt/python/pypy3.6/bin/pypy3"
-      - id: "tornado5-py35"
-        variables:
-           TOX_ENV: "tornado5-py35"
-           PYTHON_BINARY: "/opt/python/3.5/bin/python3"
       - id: "tornado5-py36"
         variables:
            TOX_ENV: "tornado5-py36"
@@ -780,18 +772,10 @@ axes:
         variables:
            TOX_ENV: "tornado5-py37"
            PYTHON_BINARY: "/opt/python/3.7/bin/python3"
-      - id: "tornado6-pypy35"
-        variables:
-          TOX_ENV: "tornado6-pypy35"
-          PYTHON_BINARY: "/opt/python/pypy3.5/bin/pypy3"
       - id: "tornado6-pypy36"
         variables:
           TOX_ENV: "tornado6-pypy36"
           PYTHON_BINARY: "/opt/python/pypy3.6/bin/pypy3"
-      - id: "tornado6-py35"
-        variables:
-           TOX_ENV: "tornado6-py35"
-           PYTHON_BINARY: "/opt/python/3.5/bin/python3"
       - id: "tornado6-py36"
         variables:
            TOX_ENV: "tornado6-py36"
@@ -816,18 +800,10 @@ axes:
         variables:
            TOX_ENV: "tornado_git-py37"
            PYTHON_BINARY: "/opt/python/3.7/bin/python3"
-      - id: "asyncio-pypy35"
-        variables:
-          TOX_ENV: "asyncio-pypy35"
-          PYTHON_BINARY: "/opt/python/pypy3.5/bin/pypy3"
       - id: "asyncio-pypy36"
         variables:
           TOX_ENV: "asyncio-pypy36"
           PYTHON_BINARY: "/opt/python/pypy3.6/bin/pypy3"
-      - id: "asyncio-py35"
-        variables:
-           TOX_ENV: "asyncio-py35"
-           PYTHON_BINARY: "/opt/python/3.5/bin/python3"
       - id: "asyncio-py36"
         variables:
            TOX_ENV: "asyncio-py36"
