+++ setup.py
@@ -3,8 +3,8 @@ from distutils.cmd import Command
 from distutils.errors import DistutilsOptionError
 from setuptools import setup
 
-if sys.version_info[:2] < (3, 5):
-    raise Exception("This version of Motor requires Python>=3.5")
+if sys.version_info[:2] < (3, 6):
+    raise Exception("This version of Motor requires Python>=3.6")
 
 classifiers = """\
 Intended Audience :: Developers
@@ -12,7 +12,6 @@ License :: OSI Approved :: Apache Software License
 Development Status :: 5 - Production/Stable
 Natural Language :: English
 Programming Language :: Python :: 3
-Programming Language :: Python :: 3.5
 Programming Language :: Python :: 3.6
 Programming Language :: Python :: 3.7
 Programming Language :: Python :: 3.8
@@ -142,7 +141,7 @@ setup(name='motor',
       author='A. Jesse Jiryu Davis',
       author_email='jesse@mongodb.com',
       url='https://github.com/mongodb/motor/',
-      python_requires='>=3.5.2',
+      python_requires='>=3.6',
       install_requires=install_requires,
       extras_require=extras_require,
       license='http://www.apache.org/licenses/LICENSE-2.0',
