+++ test/tornado_tests/test_motor_change_stream.py
@@ -23,8 +23,7 @@ from tornado.testing import gen_test
 
 from test import SkipTest, env
 from test.tornado_tests import MotorTest
-from test.py35utils import wait_until
-from test.utils import get_async_test_timeout
+from test.utils import get_async_test_timeout, wait_until
 
 
 class MotorChangeStreamTest(MotorTest):
