+++ tox.ini
@@ -6,10 +6,10 @@
 [tox]
 envlist =
     # Tornado 5 supports Python 2.7.9+ and 3.4+.
-    tornado5-{pypy35,pypy36,py35,py36,py37},
+    tornado5-{pypy36,py36,py37},
 
     # Tornado 6 supports Python 3.5+.
-    tornado6-{pypy35,pypy36,py35,py36,py37,py38,py39},
+    tornado6-{pypy36,py36,py37,py38,py39},
 
     # Test Tornado on master in a few configurations.
     tornado_git-{py36,py37},
@@ -21,7 +21,7 @@ envlist =
     py3-sphinx-doctest,
 
     # asyncio without Tornado.
-    asyncio-{pypy35,pypy36,py35,py36,py37,py38,py39},
+    asyncio-{pypy36,py36,py37,py38,py39},
 
     # Test PyMongo 3.12 because 4.x breaks some APIs
     py3-pymongo-v3.12,
@@ -39,12 +39,10 @@ passenv =
     ASYNC_TEST_TIMEOUT
 
 basepython =
-    py35: {env:PYTHON_BINARY:python3.5}
     py36: {env:PYTHON_BINARY:python3.6}
     py37,synchro37: {env:PYTHON_BINARY:python3.7}
     py38: {env:PYTHON_BINARY:python3.8}
     py39: {env:PYTHON_BINARY:python3.9}
-    pypy35: {env:PYTHON_BINARY:pypy3}
     pypy36: {env:PYTHON_BINARY:pypy3}
 
     # Default Python 3 when we don't care about minor version.
@@ -55,7 +53,7 @@ deps =
     tornado6: tornado>=6,<7
     tornado_git: git+https://github.com/tornadoweb/tornado.git
 
-    {py35,py36,py37,py38,py39}: aiohttp
+    {py36,py37,py38,py39}: aiohttp
 
     sphinx: sphinx
     sphinx: aiohttp
